Added Sidebar in a separate child component.
